feat: clean up, move test cases, reformating
