+++ waltid-services/waltid-e2e-tests/src/test/kotlin/InputDescriptorMatchingTest.kt
@@ -1,5 +1,5 @@
 import id.walt.w3c.schemes.JwsSignatureScheme
-import id.walt.issuer.issuance.IssuanceExamples
+import id.walt.issuer.issuance.openapi.issuerapi.IssuanceExamples
 import id.walt.issuer.issuance.IssuanceRequest
 import id.walt.oid4vc.util.JwtUtils
 import id.walt.webwallet.db.models.WalletCredential
