+++ waltid-services/waltid-e2e-tests/src/test/kotlin/LspPotentialWallet.kt
@@ -8,7 +8,7 @@ import id.walt.crypto.keys.KeyGenerationRequest
 import id.walt.crypto.keys.KeySerialization
 import id.walt.crypto.keys.KeyType
 import id.walt.issuer.issuance.IssuanceRequest
-import id.walt.issuer.lspPotential.LspPotentialIssuanceInterop
+import id.walt.issuer.feat.lspPotential.LspPotentialIssuanceInterop
 import id.walt.oid4vc.data.*
 import id.walt.oid4vc.requests.AuthorizationRequest
 import id.walt.sdjwt.SDField
