--- waltid-services/waltid-issuer-api/src/main/kotlin/id/walt/issuer/issuance/SignRequests.kt
@@ -1,11 +1,0 @@
-package id.walt.issuer.issuance
-
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonObject
-
-@Serializable
-data class SignRequest(
-    val issuerKey: JsonObject,
-    val vc: JsonElement,
-)
