+++ waltid-services/waltid-issuer-api/src/test/kotlin/id/walt/issuance2/NewIssuanceStub.kt
@@ -1,4 +1,4 @@
-package id.walt.issuer
+package id.walt.issuance2
 
 import id.walt.commons.logging.LoggingManager
 import id.walt.commons.logging.setups.TraceLoggingSetup
