+++ waltid-services/waltid-issuer-api/src/test/kotlin/id/walt/issuance2/Values.kt
@@ -1,4 +1,4 @@
-package id.walt.issuer
+package id.walt.issuance2
 
 object Values {
     const val VERSION = "1.SNAPSHOT"
